requirements.txt
openai
pandas
plotly
streamlit
yfinance
python-dotenv